- Ya se pueden modificar la divisa y la tasa de conversión de los documenots. - Añadida opción para imprimir albaranes y facturas de compra. - Cambiado nombrecomercial por razonsocial en clientes y proveedores. Pendiente de revisión final.

+++ controller/ventas_familias.php
@@ -63,14 +63,18 @@ class ventas_familias extends fs_controller
       }
       else if( isset($_GET['delete']) )
       {
-         $fam = new familia();
-         $fam->codfamilia = $_GET['delete'];
-         if( $fam->delete() )
+         $fam = $this->familia->get($_GET['delete']);
+         if($fam)
          {
-            $this->new_message("Familia ".$_GET['delete']." eliminada correctamente");
+            if( $fam->delete() )
+            {
+               $this->new_message("Familia ".$_GET['delete']." eliminada correctamente");
+            }
+            else
+               $this->new_error_msg("¡Imposible eliminar la familia ".$_GET['delete']."!");
          }
          else
-            $this->new_error_msg("¡Imposible eliminar la familia ".$_GET['delete']."!");
+            $this->new_error_msg("Familia ".$_GET['delete']." no encontrada.");
       }
       
       if($this->query != '')
@@ -97,7 +101,7 @@ class ventas_familias extends fs_controller
    
    private function check_familias()
    {
-      $data = $this->db->select("select * from familias where madre is not null and madre not in (select codfamilia from familias);");
+      $data = $this->db->select("SELECT * FROM familias WHERE madre IS NOT NULL AND madre NOT IN (SELECT codfamilia FROM familias);");
       if($data)
       {
          foreach($data as $d)
